backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
numpy==1.26.4
soundfile==0.12.1
faster-whisper==1.0.3
requests==2.32.3
transformers==4.46.3
torch>=2.2.0
sentencepiece==0.2.0
